requirements.txt
streamlit
google-generativeai
Pillow
gspread
oauth2client
pandas
plotly